pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "freeman-librarian"
version = "0.1.0"
description = "Document-centric world-modeling fork of Freeman."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "networkx>=3.2",
    "numpy>=1.26",
    "PyYAML>=6.0",
    "scipy>=1.11",
]

[project.optional-dependencies]
semantic = ["chromadb>=0.5"]
causal = [
    "causalml>=0.15",
    "pandas>=2.0",
    "scikit-learn>=1.3",
]
benchmark = [
    "chromadb>=0.5",
    "pandas>=2.2",
    "tqdm>=4.66",
]
dev = ["pytest>=8.0"]
mcp = ["mcp>=1.27"]
documents = [
    "pypdf>=5.0",
    "python-docx>=1.1",
]
demo = [
    "huggingface_hub>=0.30",
]

[project.scripts]
freeman = "freeman.interface.cli:main"
freeman-mcp = "freeman.mcp.server:main"
freeman-librarian = "freeman_librarian.interface.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["freeman*", "freeman_librarian*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
